src/css_path.rs: move code into `mini_selector`
