+++ src/lib.rs
@@ -16,11 +16,11 @@
 #![doc= include_str!("../Examples.md")]
 
 mod css;
-mod css_path;
 mod document;
 mod dom_tree;
 mod entities;
 mod matcher;
+mod mini_selector;
 mod node;
 mod selection;
 mod serializing;
@@ -29,6 +29,8 @@ pub use document::Document;
 pub use dom_tree::Tree;
 pub use dom_tree::TreeNodeOps;
 pub use matcher::Matcher;
+pub use mini_selector::MiniSelector;
+pub use mini_selector::{parse_selector_list, parse_single_selector};
 #[doc(hidden)]
 pub use node::SerializableNodeRef;
 pub use node::{Element, Node, NodeData, NodeId, NodeIdProver, NodeRef};
