+++ tests/node-traversal.rs
@@ -456,7 +456,6 @@ fn test_doc_format_md_table() {
     assert_eq!(text.as_ref(), expected);
 }
 
-
 #[cfg_attr(not(target_arch = "wasm32"), test)]
 #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
 fn test_node_find_by_css() {
@@ -473,7 +472,6 @@ fn test_node_find_by_css() {
         .collect();
 
     assert_eq!(got_ids, expected_ids);
-    
 
     let a_sel = doc.select("a");
     let expected_ids: Vec<dom_query::NodeId> = a_sel.nodes().iter().map(|n| n.id).collect();
@@ -485,7 +483,6 @@ fn test_node_find_by_css() {
     let len_sel_ne = doc.select("body td p").length();
     assert_eq!(len_sel_ne, 0)
 }
-
 
 #[cfg_attr(not(target_arch = "wasm32"), test)]
 #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
@@ -499,13 +496,8 @@ fn test_node_find_by_css_combinator() {
         let expected_ids: Vec<dom_query::NodeId> = a_sel.nodes().iter().map(|n| n.id).collect();
 
         let root = doc.root();
-        let got_ids: Vec<dom_query::NodeId> = root
-            .find_by_css(sel)
-            .iter()
-            .map(|n| n.id)
-            .collect();
+        let got_ids: Vec<dom_query::NodeId> = root.find_by_css(sel).iter().map(|n| n.id).collect();
 
         assert_eq!(got_ids, expected_ids);
     }
-    
 }
